+++ R/afs_capewide_pup.R
@@ -12,46 +12,55 @@
 #'   `cwp_total_by_loc()`. This argument is intended to minimize work if a user
 #'   needs both the output of `cwp_total_by_loc()` and `cwp_total()`
 #'
-#' @details `cwp_loc_agg()`: Aggregate pup counts and related info across
+#' @details Function-specific details:
+#'  * `cwp_loc_agg()`: Aggregate pup counts and related info across
 #'   locations, and by season/observer. This function is intended for situations
 #'   where a user has filtered CWP data for a specific subset of locations and
 #'   wants to summarize the counts across all of these locations, effectively
 #'   treating all of the locations as one location.
 #'
-#'   `cwp_review()`: Summarize pup counts and ancillary information for review
+#'   * `cwp_review()`: Summarize pup counts and ancillary information for review
 #'   after grouping by season_name and location. `loc.agg == TRUE` then all
 #'   counts will be summed by observer across locations, i.e., `x` will be
 #'   passed to `cwp_loc_agg()`, before being summarized. The output data frame
 #'   is various pasted conglomerations that are intended for review to help
 #'   determine if recounts need to happen.
 #'
-#'   `cwp_total_by_loc()`: group by season_name and location; return count mean,
-#'   variance, SD, and min date. Calculating the variance and SD follow
-#'   CS/AMLR-specific rules about when it must be NA based on historical
-#'   knowledge about past survey data; see the 'assumptions' section for
+#'   * `cwp_total_by_loc()`: group by season_name and location; return count mean,
+#'   variance, SD, and min date. Calculating the variance and SD follow Cape
+#'   Shirref/AMLR-specific rules about when it must be NA based on historical
+#'   knowledge about past survey data. See the 'assumptions' section for
 #'   details. census_date_min is the earliest census date for that
-#'   season/location group.
+#'   season/location group. Records with
 #'
-#'   `cwp_total()`: after running cwp_total_by_loc, group by season name and
+#'   * `cwp_total()`: after running cwp_total_by_loc, group by season name and
 #'   return counts and standard deviations
 #'
 #'   Note that none of these functions do any rounding. To round values, for
-#'   instance to get a whole number for pup counts, use either [base::round()]
-#'   or [amlrDatabases::round_logical()]
+#'   instance to get a whole number for pup counts, users may use functions such
+#'   as [base::round()] or [amlrDatabases::round_logical()]
 #'
-#'   These functions make several assumptions specific to US AMLR Pinniped Data:
-#'   * Assumes that the research program is unique by season
+#'   The cwp functions make several assumptions specific to U.S. AMLR Pinniped
+#'   Data:
+#'   * The research program values are unique when considered by season
 #'   * Study beaches (locations: Copihue, Maderas, Cachorros, Chungungo)
 #'   censused between 2008-07-01 and 2011-07-01 are single counts and thus do
 #'   not have variance values.
-#'   * All counts before July 2008 (i.e., before the 2008/09 season) do not have variance values.
+#'   * All counts before July 2008 (i.e., before the 2008/09 season)
+#'   do not have variance values.
 #'
-#' @return An ungrouped data frame with the summarized AFS Capewide Pup Census
-#'   data. The summary types are described in 'Details'.
+#' @return An ungrouped object of the same tpye as `x`, with the summarized AFS
+#'   Capewide Pup Census data. The summary options are described in 'Details'.
+#'
+#'   For all functions except `cwp_total()`: if 'census_afs_capewide_pup_sort'
+#'   exists as a column in `x`, then the output data frame will also be sorted
+#'   by 'census_afs_capewide_pup_sort' (after being sorted first by the column
+#'   'season_name').
 #'
 #' @examplesIf FALSE
-#'   # Not run; examples only will run if on SWFSC network
-#'   x <- tbl_vCensus_AFS_Capewide_Pup(amlr_dbConnect("***REMOVED***_Test"))
+#'   # Not run; examples only will run if `con` is a valid database connection
+#'   con <- amlr_dbConnect("***REMOVED***")
+#'   x <- tbl_vCensus_AFS_Capewide_Pup(con)
 #'
 #'   x.201617 <- x[x$season_name == "2016/17", ]
 #'   cwp_review(x.201617)
@@ -70,7 +79,7 @@
 #' @export
 cwp_loc_agg <- function(x) {
   columns.names <- c(
-    "season_name", "census_afs_capewide_pup_sort", "location", "census_date",
+    "season_name", "location", "census_date",
     "observer", "census_notes", "exclude_count",
     "pup_count", "pup_live_count", "pup_dead_count"
   )
@@ -78,18 +87,16 @@ cwp_loc_agg <- function(x) {
     all(c(columns.names %in% names(x)))
   )
 
-  if (any(x$exclude_count)) {
-    warning("x contains some records with exclude_count == TRUE; ",
-            "note these records will be removed to aggreagate across location",
-            immediate. = TRUE)
-  }
+  sort.syms <- syms(intersect(c("census_afs_capewide_pup_sort"), names(x)))
 
-  x %>%
-    arrange(census_afs_capewide_pup_sort) %>%
-    filter(!exclude_count) %>%
+  x <- cwp_exclude_count(x)
+
+  x.out <- x %>%
+    arrange(!!!sort.syms) %>%
     group_by(season_name, observer) %>%
     summarise(across(c(pup_count, pup_live_count, pup_dead_count),
-                     sum, na.rm = TRUE, .names = "{.col}"),
+                     \(x) sum(x, na.rm = TRUE),
+                     .names = "{.col}"),
               location = paste(unique(location), collapse = ", "),
               census_date_min = min(census_date),
               exclude_count = unique(exclude_count),
@@ -98,9 +105,14 @@ cwp_loc_agg <- function(x) {
                 NA_character_,
                 paste(as.character(na.omit(census_notes)), collapse = "; ")),
               research_program = unique(research_program),
-              .groups = "drop") %>%
-    mutate(census_afs_capewide_pup_sort = 1,
-           .after = season_name)
+              .groups = "drop")
+
+  if (length(sort.syms) > 0) {
+    x.out %>%
+      mutate(census_afs_capewide_pup_sort = 1, .after = season_name)
+  } else {
+    x.out
+  }
 }
 
 
@@ -109,7 +121,7 @@ cwp_loc_agg <- function(x) {
 #' @export
 cwp_review <- function(x, loc.agg = FALSE) {
   columns.names <- c(
-    "season_name", "census_afs_capewide_pup_sort", "location", "census_date",
+    "season_name","location", "census_date",
     "observer", "census_notes", "exclude_count",
     "pup_count", "pup_live_count", "pup_dead_count"
   )
@@ -117,11 +129,13 @@ cwp_review <- function(x, loc.agg = FALSE) {
     all(c(columns.names %in% names(x)))
   )
 
+  sort.syms <- syms(intersect(c("census_afs_capewide_pup_sort"), names(x)))
+
   # Aggregate across locations, if specified
   if (loc.agg) x <- cwp_loc_agg(x) %>% rename(census_date = census_date_min)
 
   x %>%
-    group_by(season_name, census_afs_capewide_pup_sort, location) %>%
+    group_by(season_name, !!!sort.syms, location) %>%
     arrange(observer) %>% #so that collapsed data are always in the same order
     summarise(n_records = n(),
               count_mean = mean(pup_count),
@@ -135,28 +149,27 @@ cwp_review <- function(x, loc.agg = FALSE) {
               counts_live = paste(pup_live_count, collapse = "; "),
               counts_dead = paste(pup_dead_count, collapse = "; "),
               .groups = "drop") %>%
-    relocate(census_afs_capewide_pup_sort, .after = last_col())
+    relocate(!!!sort.syms, .after = last_col())
 }
 
 #' @name afs_capewide_pup
 #' @export
 cwp_total_by_loc <- function(x, loc.agg = FALSE) {
   columns.names <- c(
-    "season_name", "census_afs_capewide_pup_sort", "location",
-    "census_date", "pup_count", "research_program"
+    "season_name", "location",
+    "census_date", "pup_count", "research_program", "exclude_count"
   )
   stopifnot(all(c(columns.names %in% names(x))))
-  if (any(x$exclude_count)) {
-    warning("x contains some records with exclude_count == TRUE; ",
-            "note these records will be removed to summarize totals by location",
-            immediate. = TRUE)
-  }
+
+  sort.syms <- syms(intersect(c("census_afs_capewide_pup_sort"), names(x)))
+
+  x <- cwp_exclude_count(x)
 
   # Aggregate across locations, if specified
   if (loc.agg) x <- cwp_loc_agg(x) %>% rename(census_date = census_date_min)
 
   x %>%
-    group_by(season_name, census_afs_capewide_pup_sort, location) %>%
+    group_by(season_name, !!!sort.syms, location) %>%
     summarise(num_records = n(),
               count_loc_mean = mean(pup_count),
               count_loc_var = var(pup_count),
@@ -173,19 +186,20 @@ cwp_total_by_loc <- function(x, loc.agg = FALSE) {
            ),
            count_loc_sd = sqrt(count_loc_var)) %>%
     relocate(count_loc_sd, .before = count_loc_var) %>%
-    select(-c(census_afs_capewide_pup_sort, study_beach_count))
+    select(-c(study_beach_count))
 }
 
 #' @name afs_capewide_pup
 #' @export
 cwp_total <- function(x, x.byloc = FALSE, loc.agg = FALSE) {
   x.df.byloc <- if (x.byloc) {
-    stopifnot(identical(
-      c("season_name", "location",
-        "num_records", "count_loc_mean", "count_loc_sd", "count_loc_var",
-        "census_date_min", "research_program"),
-      names(x)
-    ))
+    loc.columns.names <- c(
+      "season_name", "location",
+      "num_records", "count_loc_mean", "count_loc_sd", "count_loc_var",
+      "census_date_min", "research_program"
+    )
+    stopifnot(all(loc.columns.names %in% names(x)))
+
     x
   } else {
     cwp_total_by_loc(x, loc.agg = loc.agg)
@@ -200,4 +214,22 @@ cwp_total <- function(x, x.byloc = FALSE, loc.agg = FALSE) {
               research_program = unique(research_program),
               .groups = "drop") %>%
     select(-count_var)
+}
+
+
+
+# Verbosely remove any records with an exclude_count flag
+cwp_exclude_count <- function(x) {
+  if (any(x$exclude_count)) {
+    warning("x contains ", sum(x$exclude_count),
+            " record(s) with exclude_count == TRUE; ",
+            "note that these records will be filtered out ",
+            "for the specified processing",
+            immediate. = TRUE)
+
+    x %>% filter(exclude_count == 0)
+
+  } else {
+    x
+  }
 }