+++ NEWS.md
@@ -8,13 +8,13 @@
 
 * Updated various tab functionality, especially AFS DCC and cape-wide pup census
 
-* Updated `afs_capewide_pup` functions: small internal changes; renamed functions including changing to 'cwp_' prefix; added `cwp_loc_agg` to aggregate across locations in a consistent fashion.
+* Updated `afs_capewide_pup` functions: renamed functions including changing to 'cwp_' prefix; added `cwp_loc_agg` to aggregate across locations in a consistent fashion; bug fixes (#10)
 
 * Extracted `dcc` functions from Shiny app and made them generalized functions for processing DCC download data. 
 
 * Added DCC sample data [#5]
 
-* Added tests for `dcc`, `csphoc`, and ... functions [e.g., #6]
+* Added tests for `dcc`, `csphoc`, and `cwp` functions [e.g., #6]
 
 
 # amlrPinnipeds 0.5.0